Added test file for overviewpage, added tests for claimstatuspage and claimstatuspagecontent
